StatsPanel
the Stats Panel is added
