+++ src/display/Frame.java
@@ -1,11 +1,15 @@
 package display;
+
+import java.awt.GridLayout;
+
 import javax.swing.JFrame;
 
 public class Frame extends JFrame {
-    Panel panel = new Panel();
 
     Frame() {
+    	this.setLayout(new GridLayout(1, 2));
     	this.add(new Panel());
+    	this.add(new StatsPanel());
 		this.setTitle("Tertrix");
 		this.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 		this.setResizable(false);
